requirements.txt
streamlit
pillow
python-dateutil
regex
